refactor(reflect): keep only SKILL.md at the skill's top level
The four bundled files sat beside SKILL.md. Move them under
skills/reflect/resources/ so the skill directory presents one entry
point and its supporting files are grouped beneath it.

Paths updated at every reference: the two node invocations in the skill
body, the catalog entry in docs/skills.md, the test imports, and the
usage line in write-target.mjs. The scripts' own usage strings print
basenames and are unaffected. docs/cross-host-portability.md lists
bundled scripts by basename across all skills, so it stays accurate.
