+++ skills/reflect/resources/resolve-transcript.d.mts
