+++ skills/reflect/resources/write-target.d.mts
